repository: Return error when trying to prune objects in pool repos
The PruneUnreachableObjects RPC only purpose is to prune unreachable
objects. While this should theoretically be a relatively safe operation
for normal repositories, it definitiely isn't for object pools as any
pool members may have deduplicated their objects so that they rely on
the objects to keep existing in the pool repository. The concept of
reachability does not really play any role here: once an object is in an
object pool it must stay there.

The PruneUnreachableObjects RPC does not have any safeguards for this
though. Let's put one into place and return an error in case we are
unexpectedly executed on an object pool.

Changelog: changed