repository: Consider cruft packs in PruneUnreachableObjects
When pruning objects via PruneUnreachableObjects, then we execute
git-prune(1) with a reduced grace period of 30 minutes. This only works
to prune loose objects though, and will never delete objects part of a
cruft pack. This means that as we're moving unreachable objects into
cruft packs, that those objects will in fact not get pruned as expected
anymore.

Fix this bug by repacking the repository with cruft packs and a reduced
expiry time.

Changelog: fixed