Configure pack-objects min occurrences via env var
Inject the MinOccurrences middleware into the streamcache call stack
and make it configurable via the
GITALY_PACK_OBJECTS_CACHE_MIN_OCCURRENCES environment variable.

This is a temporary configuration mechanism that will allow us to
validate the MinOccurrences middleware on GitLab.com quicker. If it
works as expected we will replace the env var with normal config.toml
configuration logic.