statushandler: Test for context errors with gRPC errors
The statushandler middlerware is responsible for converting errors into
either a `DeadlineExceeded` or `Canceled` error when the RPC has been
canceled via the context. This functionality recently broke when we
changed behaviour of the `structerr` package around how gRPC errors are
propagated in 076d3b6bb (Merge branch 'pks-structerr-unknown-by-default'
into 'master', 2023-03-17).

The fact that things could break indicates that we are lacking test
coverage for cases where the RPC implementation returns an error with a
specific gRPC error code which should then be overridden via the context
error. Add tests that would've cought the regression.