Remove streamcache index cleanup goroutine
This simplifies streamcache by getting rid of one of its cleanup
goroutines. After this change there is one goroutine left, that does
file walks and deletes cache files. This is so slow that it makes
sense to do it in a goroutine. Kicking keys out of the in memory index
now happens during Fetch calls.

Changelog: other