cli: Add gitaly-bench coordinator
Currently we use Ansible to run benchmarks against Gitaly, but this
requires us to run benchmarks serially. Ansible cannot divide a group of
tasks between hosts, so we're limited to a single Gitaly host. The goal
of this tool is to enable us to create `n` Gitaly hosts and run
benchmarks in parallel against them.

This initial commit creates the `Coordinator`, which is run on the
Gitaly hosts being benchmarked and is responsible for starting and
stopping Gitaly when commanded by the client machine. The `Coordinator`
listens on a separate TCP port from Gitaly, running until an error
occurs or told to exit by the client.

In the future the ability to capture Gitaly's logs and run profiling
will be added.