cli: Add `client` command to `gitaly-bench`
Add the `client` subcommand to `gitaly-bench`, which is run on the
host that will run benchmarking requests against Gitaly. This sends
JSON-encoded commands to the `Coordinator` via TCP, ordering it to start
and stop Gitaly for benchmarking.

Details of the rpcs to benchmark and the specific queries to run are
stored on-disk and located at runtime.