cmd: Add `gitaly-bench` binary
Add the new `gitaly-bench` binary for coordinating starting and stopping
Gitaly for benchmarking.

The tests at this level do not attempt to verify anything other than the
basic CLI output as `gitaly-bench` requires a benchmarking environment
to run successfully. In particular we need to have a `gitaly` service
available in `systemd` for invocations to succeed.