trace2: Add Trace2 pack-objects metrics exporter to Git command factory
This commit integrates Trace2 tracing expoter from some prior commits to
Git comand factory. This hook is controlled by feature flag
`gitaly_export_trace2_pack_objects_metrics`. After this flag is on, this
hook is unlikely to affect the overall performance because the rate of
this command is not high.