lstree: Move package to localrepo
The lstree package contains a helper for reading a tree. There is no
reason why this code needs to be in its own package. In fact, this makes
it harder when for instance, we might want to add code in the
`localrepo` package that calls `ListEntries()` since it would result in
a dependency cycle because `lstree` depends on `localrepo`.

Move the code under `lstree` under `localrepo`, since that is the one
stop shop for operations on local repositories.