gitaly/maintenance: Enable testing with SHA256 object format
Enable testing with the SHA256 object format in Gitaly's maintenance
tests.

Only a single change is required: we used to call git-init(1) manually
to initialize repositories for one of our tests, which of course would
not be aware of the default object format. Convert these tests to use
`gittest.CreateRepository()` instead.