cleanup: Parse object map with detected object hash
We're using hardcoded object ID widths in the cleanup service in order
to parse the object map that maps old to new object IDs. This will of
course break when using repositories with the SHA256 object format.

Convert the code to detect the object hash and use `ValidateHex()` to
verify that the mappings are indeed valid.