git/housekeeping: Refactor repacking objects to use repacking strategy
We support different strategies to repack a repository: incremental,
full and full with cruft packs. These are mutually exclusive and work
differently under the hood. We're also about to add a fourth mode with
geometric repacking that makes the infrastructure even more complex.

Let's refactor the code to make explicit that these are different
strategies and to make it easier to add new ones or remove old ones.