featureflag: Remove `tx_restore_custom_hooks`
The `tx_restore_custom_hooks` feature flag enables transactionality and
atomicity when setting custom Git hooks for a repository through the
`SetCustomHooks` RPC. With the release of GitLab 15.10 this flag was
default enabled and everything appears to be working well. This change
removes the feature flag along with the call sites checking for it.