golangci-lint: Use `tenv` linter instead of custom forbidigo rules
Since the introduction of the `t.Setenv()` helpers in Go 1.17 it has
become discourages to use `os.Setenv()` in tests directly. This is
because `t.Setenv()` has additional safeguards that cause the test to
fail if called in a parallel test as setting environment variables is
not thread-safe.

We enforce this rule via a set of custom forbidigo rules. But since
adding those rules, a new `tenv` linter has been added to golangci-lint
that supports exactly that usecase.

Use the new `tenv` linter and drop our own custom rules.