git/stats: Convert multi-pack-index packfile count to uint64
The multi-pack-index info contains the number of packfiles tracked by
it. As the Git file format uses a 32-bit unsigned integer for this field
we also store it as a `uint32`. All the other fields we typically store
though are tracked via `uint64`, which makes it somewhat uncomfortable
to use the field without putting casts at many of the callsites.

Convert the field to `uint64` to avoid casts.