Dangerfile: Add group labels
When an MR is created, we can automatically add group::gitaly based on
the mr author. Additionally, we can inherit gitaly::cluster and
gitaly::git labels as well from the linked issue.

e5e9234f3 (danger: Remove tracking of team members, 2021-12-17) had
removed this because it was a static list that would easily grow stale.
Now that we are pulling from the group, that mitigates the staleness
risk.