git/stats: Count cruft and kept packfiles into all packfiles
The documentation we have in place for both the count and the size of
packfiles mentions that this counter also includes stale packfiles. This
behaviour is also the intended one so that callers can easily tell how
big all packfiles together are even when additional categories of them
are added over time. Furthermore, this also matches the way we count
loose objects.

But even though it's quite clear that this is the intended behaviour it
is not the actual behaviour: neither cruft packs nor kept packs are
counted into the overall number of packfiles.

Fix this inconsistency.

Changelog: fixed