repository: Refactor GetArchive tests for exact match of contents
In our GetArchive tests we're executing tar(1) and unzip(1) in order to
verify contents of the generated archive. This is inflexible as we don't
want to parse their outputs and creates a dependency on system-provided
binaries.

Refactor the `compressedFileContents()` helper function to instead use
Go's own packages to parse the archives. This allows us to easily have
it return a map of paths and their respective contents, which we'll use
in a subsequent commit to make the tests stricter.