repository: Refactor GetArchive tests to use exact matching
We're using seed repositories in the GetArchive tests. While discouraged
for SHA256 compatibility, it also means that it's hard to use exact
matches for the contents of the generated archive as the list of files
is comparatively large.

Refactor the test to generate test data at runtime. This gives us full
control over the actual files and thus allows us to convert the tests to
do exact matches for expected archive contents.