repository: Deseed remaining tests for GetArchive
Deseed remaining tests for GetArchive so that we can start testing with
SHA256.

While at it, stop skipping one of the tests when running with Praefect.
The reason for skipping is that we use an intercepted command factory,
which in turns breaks creating repositories via the CreateRepository
RPC. But that's not true, as we only intercept calls to git-archive(1).