catfile: Expose object format via ObjectInfo
While both the catfile ObjectContentReader and ObjectInfoReader know
about the object format that is currently in use, that information is
not exposed to the user of those interfaces.

Expose the object format via the `ObjectInfo` structure so that callers
can use this information. Note that we explicitly don't embed the full
`git.ObjectHash` structure, but only the name of the object format. This
is done to keep the size of the structure as small as possible as we
often process thousands of objects via the gitpipe package.