repository: Move requests for GetArchive tests into testcase
The testcases for the GetArchive RPC replicate the structure of the
request itself only to fill in the request with that self-implemented
structure. This is quite roundabout, duplicates logic and is not as
flexible as if we just had the request in the testcase itself.

Refactor the code to expect a GetArchiveRequest in the testcases instead
of replicating its structure.