Refator PackObjectsLimiting configs
This commit adds MaxQueueLength field to make it consistent with other
limiters. As the concurrency limiter is new, we don't want to add a
support for such configs to Omnibus/K8s helm chart too soon. As a
result, the configs can be fed from the environment variables.

Besides, this commit moves the default PackObjectsLimiting
configurations to `internal/config/config.go`. That file is the
centralized location to store all Gitaly configurations as well as
default setting.