git/housekeeping: Deduplicate common repack configuration
Some fields of the `RepackObjectsConfig` are always the same no matter
what repack strategy we're using, but we still duplicate them across the
different branches. This is somewhat awkward and makes it easy to miss
this common configuration when adding new heuristics.

Refactor the code so that the common parts are deduplicated.