git/housekeeping: Fix missing pack-objects strategy metric
Part of the metrics we expose in OptimizeRepository is the way we have
packed objects. This is implemented via a switch over the strategy used
to repack objects, which is fragile as it needs to be updated whenever
we implement a new strategy. In fact, the switch is already missing the
"geometric" strategy.

Now that the RepackObjectsStrategy is backed by a string though we can
simplify the code and avoid switching altogether. Do so.

Changelog: fixed