Enable trace2 pack-objects metrics for git-pack-objects
In some prior commits, Gitaly allows hooking into pack-objects command
to extract relative metrics. The current implementation enables that
hook for all Git commands, not just git-pack-objects. While it does not
disrupt the activities of other commands, enabling it there is a waste
of resources. It's very fortunate for us that it's behind a feature
flag.

This commit fixes this situation by filtering the sub command that can
enable this hook.