Replace IsPositive with GreaterOrEqual
The check IsPositive is not that useful when we have other
comparison function. Also, it confuses as 0 is not a positive
number. It is replaced with usage of the GreaterOrEqual where
value for comparison is set to 0. The IsPositive is removed
and all call sites are changed to use GreaterOrEqual.