git/housekeeping: Add strategy to repack with unreachable objetcs
Even with the introduction of geometric repacking we'll need to keep on
alternating between full and geometric repacks in object pools. This is
required so that we can "freshen" delta islands every now and then as
geometric repacks do not consider delta islands at all.

That being said, we don't have a proper strategy to perform full repacks
in object pools:

    - Using cruft packs is quite pointless as we don't want to ever
      delete unreachable objects.

    - Using our usual strategy to perform a full repack explodes
      unreachable objects into loose objects, which is inefficient.
      Furthermore, the next geometric repack would then again soak up
      all new loose objects, which is quite inefficient.

So we need a better strategy to handle unreachable objects in pool
repositories during a full repack.

Luckily, git-repack(1) provides the `--keep-unreachable` option which
will cause it to just append all already-packed unreachable objects to
the new packfile. While it would be preferable if it _also_ included
loose unreachable objects, there is currently no way to do so.

Implement this repack mode so that we can use it for our new strategy.