git: Refactor object hash detection
Object hash detection is implemented via `git.DetectObjectHash()`, which
accepts a `git.RepositoryExecutor` as input to detect the object hash
for. This is unpractical as we don't always have an executor around.
Furthermore, it is also roundabout: the `RepositoryExecutor` interface
already has as `ObjectHash()` function available which is supposed to
detect the object format, so there wouldn't be any point in calling out
to `DetectObjectHash()` in the first place.

Refactor the code to accept a `git.CommandFactory` and repository
instead.