git/housekeeping: Make test for cleaning up worktrees more robust
The `cleanDisconnectedWorktrees()` function will delete worktrees that
are disconnected, which means a worktree that still has its metadata in
`$GIT_DIR/worktrees` but where the actual worktree does not exist
anymore.

In order to avoid needless busywork, we only execute the corresponding
Git command in case there actually are worktrees in that directory. We
verify that this works in our tests by using an intercepting Git command
factory that fails with a specific exit code once executed.

While this works, it is both fragile and doesn't really verify our
expectations. So let's convert the function to use a counting Git
command factory instead so that we can also observe the effect and
verify that git-worktree(1) is really only executed once.

This prepares for an upcoming change where we start to execute another
Git command to figure out the object format when setting up the hooks
payload.