remote: Don't needlessly use hooks in FindRemoteRootRef
The FindRemoteRootRef RPC executes `git remote show` in order to figure
out what the default references of a remote repository is. This RPC does
not change anything in the repository it's running in -- in fact, it
would even work without any repository at all. But we still set up the
reference-transaction hook needlessly.

Stop doing so and instead disable hooks for the command.