gitaly/hook: Fix detection of force-deletion-only changes with SHA256
The reference-transaction hook needs to be able to detect changes which
only consists of force deletions. To do so, it checks whether all
updates start with the force-deletion prefix of two zero object IDs.
This prefix is currently using the SHA1 object hash though, which breaks
compatibility with the SHA256 object format.

Use the new object format embedded in the hooks payload to look up the
actual object format used by the repository. Fix the tests to not use
hardcoded values, either.