gitaly/hook: Fix parsing input for SHA256 repositories
The update hook verifies that the old and new value it receives as input
are properly formatted object hashes. Right now, it assumes that the
object format that the repository uses is SHA1, which of course breaks
compatibility with SHA256 repositories. While tests of this hook do in
fact already test with SHA256, they have been broken due to using hard
coded values as object IDs.

Use the new object format embedded in the hooks payload to look up the
actual oject format used by the repository. Fix the tests to use actual
object IDs.