git/housekeeping: Don't loosen objects when eagerly packing object pools
When eagerly repacking object pools we use a strategy that repacks all
reachable objects into one pack, which is what we want. But we also
explode unreachable objects into loose objects, which does not make any
sense at all as we wouldn't ever delete these unreachable objects
anyway.

This really is a theoretical problem, only: there exist no callers that
invoke the eager optimization strategy on object pools. But let's fix it
regardless of that and instead use a strategy where unreachable objects
will instead be appended to the packfile.

Changelog: fixed