Revert "git/housekeeping: Introduce geometric repacking strategy"
Revert introduction of the geometric repacking strategy via 12bde1a93
(git/housekeeping: Introduce geometric repacking strategy, 2023-03-30),
which is causing the following failure in Rails' tests:

    1) Backup::Repositories#restore restoring object pools schedules restoring of the pool
     Failure/Error: raise Gitlab::Git::CommandError, e

     # GRPC::Internal:
     #   13:could not repack: exit status 128. debug_error_string:{"created":"@1680266558.837598574","description":"Error received from peer unix:tmp/tests/gitaly/praefect.socket","file":"src/core/lib/surface/call.cc","file_line":1063,"grpc_message":"could not repack: exit status 128","grpc_status":13}
     #   ./lib/gitlab/gitaly_client.rb:185:in `execute'

The failure mode is kind of puzzling: git-repack(1) complains about a
packfile not existing even though it is clearly present in the
repository. Potentially we are hitting a bug in Git here.